+++ serve.go
@@ -62,7 +62,6 @@ func (c *ServeCommand) Init(ctx context.Context, globals *Globals, files fs.Read
 			Index:           globals.Elastic.Index,
 			Schema:          globals.Postgres.Schema,
 			Title:           c.Title,
-			SizeField:       globals.Elastic.SizeField,
 			store:           nil,
 			coordinator:     nil,
 			storage:         nil,
@@ -84,7 +83,6 @@ func (c *ServeCommand) Init(ctx context.Context, globals *Globals, files fs.Read
 			site.Index = globals.Elastic.Index
 			site.Schema = globals.Postgres.Schema
 			site.Title = c.Title
-			site.SizeField = globals.Elastic.SizeField
 		}
 	}
 
@@ -120,7 +118,7 @@ func (c *ServeCommand) Init(ctx context.Context, globals *Globals, files fs.Read
 		siteCtx := context.WithValue(ctx, requestIDContextKey, "serve")
 		siteCtx = context.WithValue(siteCtx, schemaContextKey, site.Schema)
 
-		store, coordinator, storage, esProcessor, errE := es.InitForSite(siteCtx, globals.Logger, dbpool, esClient, site.Schema, site.Index, site.SizeField) //nolint:govet
+		store, coordinator, storage, esProcessor, errE := es.InitForSite(siteCtx, globals.Logger, dbpool, esClient, site.Schema, site.Index) //nolint:govet
 		if errE != nil {
 			return nil, nil, errE
 		}